fix pipage, push relabel

+++ common/utils/common_util.cpp
@@ -1,23 +1,23 @@
 #include "common_util.h"
 
-bool over(float value, float boundary) {
+bool over(double value, double boundary) {
 	return (value + ERROR_TOLERANCE >= boundary);
 }
 
 
-bool exceed(float value, float boundary) {
+bool exceed(double value, double boundary) {
 	return (value + ERROR_TOLERANCE > boundary);
 }
 
-bool roughly_equal(float value, float boundary) {
+bool roughly_equal(double value, double boundary) {
 	bool lower_equal = (value + ERROR_TOLERANCE >= boundary && value < boundary);
 	bool upper_equal = (value > boundary && value - ERROR_TOLERANCE <= boundary);
 	return value == boundary || lower_equal || upper_equal;
 }
 
 // TODO : check if it is correct
-float** getMergedGraph(int s, int t, int n, float** inputGraph) {
-	float** mergedGraph = create2dArr<float>(n - 1, 0.0f);
+double** getMergedGraph(int s, int t, int n, double** inputGraph) {
+	double** mergedGraph = create2dArr<double>(n - 1, 0.0f);
 
 	int newRow = 1;
 	for (int originalRow = 0; originalRow < n; originalRow++) {
